Created temperature schemas

+++ app/models/temperature_records.py
@@ -10,5 +10,5 @@ temperature_records_table = sqlalchemy.Table(
     sqlalchemy.Column("id", sqlalchemy.Integer, primary_key=True),
     sqlalchemy.Column("location_id", sqlalchemy.ForeignKey(locations_table.c.id)),
     sqlalchemy.Column("date", sqlalchemy.DateTime()),
-    sqlalchemy.Column("temperature", sqlalchemy.Float)
+    sqlalchemy.Column("temperature", sqlalchemy.Integer)
 )
